WEBAPP-126: Move Layout components to separate module
