+++ src/modules/layout/assets/integreat-app-logo.png
